@@ -11,7 +11,10 @@
     <packaging>jar</packaging>
 
     <properties>
-        <java.version>17</java.version>
+        <java.version>21</java.version>
+        <maven.compiler.source>21</maven.compiler.source>
+        <maven.compiler.target>21</maven.compiler.target>
+        <maven.compiler.release></maven.compiler.release>
         <vaadin.version>24.8.3</vaadin.version>
         <archunit.version>1.4.1</archunit.version>
     </properties>
