@@ -48,19 +48,19 @@ public class CActivitiesView extends CProjectAwareMDPage<CActivity> {
 		panel = new CPanelActivityDescription(getCurrentEntity(), getBinder(),
 			(CActivityService) entityService);
 		addAccordionPanel(panel);
-		panel = new CPanelActivityProject(getCurrentEntity(), getBinder(),
+		panel = new CPanelActivityStatusPriority(getCurrentEntity(), getBinder(),
 			(CActivityService) entityService);
 		addAccordionPanel(panel);
 		panel = new CPanelActivityResourceManagement(getCurrentEntity(), getBinder(),
-			(CActivityService) entityService);
-		addAccordionPanel(panel);
-		panel = new CPanelActivityStatusPriority(getCurrentEntity(), getBinder(),
 			(CActivityService) entityService);
 		addAccordionPanel(panel);
 		panel = new CPanelActivityTimeTracking(getCurrentEntity(), getBinder(),
 			(CActivityService) entityService);
 		addAccordionPanel(panel);
 		panel = new CPanelActivityHierarchy(getCurrentEntity(), getBinder(),
+			(CActivityService) entityService);
+		addAccordionPanel(panel);
+		panel = new CPanelActivityProject(getCurrentEntity(), getBinder(),
 			(CActivityService) entityService);
 		addAccordionPanel(panel);
 		panel = new CPanelActivityBudgetManagement(getCurrentEntity(), getBinder(),
