@@ -7,138 +7,133 @@ import jakarta.persistence.AttributeOverride;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
-import jakarta.validation.constraints.Size;
 import tech.derbent.abstracts.annotations.MetaData;
 import tech.derbent.abstracts.domains.CTypeEntity;
+import tech.derbent.projects.domain.CProject;
 
 /**
- * CCommentPriority - Domain entity representing comment priority levels.
- * Layer: Domain (MVC)
- * 
- * Inherits from CTypeEntity to provide type functionality for comment priorities.
- * Supports priority-based comment management with levels like LOW, NORMAL, HIGH, URGENT.
+ * CCommentPriority - Domain entity representing comment priority levels. Provides
+ * priority classification for comments to support workflow and escalation management.
+ * Layer: Domain (MVC) Standard priority levels: URGENT, HIGH, NORMAL, LOW, INFO
+ * @author Derbent Team
+ * @since 1.0
  */
 @Entity
-@Table(name = "ccommentpriority")
-@AttributeOverride(name = "id", column = @Column(name = "ccommentpriority_id"))
+@Table (name = "ccommentpriority")
+@AttributeOverride (name = "id", column = @Column (name = "ccommentpriority_id"))
 public class CCommentPriority extends CTypeEntity {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(CCommentPriority.class);
+	private static final Logger LOGGER = LoggerFactory.getLogger(CCommentPriority.class);
 
-    @Column(name = "priority_level", nullable = false)
-    @MetaData(
-        displayName = "Priority Level", required = true, readOnly = false,
-        defaultValue = "2", description = "Numeric priority level (1=Highest, 4=Lowest)",
-        hidden = false, order = 3
-    )
-    private Integer priorityLevel = 2; // Default to NORMAL priority
+	/**
+	 * Priority level for the comment (1=Highest, 4=Lowest).
+	 */
+	@Column (name = "priority_level", nullable = false, length = 20)
+	@MetaData (
+		displayName = "Priority Level", required = false, readOnly = false,
+		defaultValue = "LOW", description = "Severity level of the risk", hidden = false,
+		order = 2, useRadioButtons = false
+	)
+	Integer PriorityLevel;
 
-    @Column(name = "color", nullable = true, length = 7)
-    @Size(max = 7)
-    @MetaData(
-        displayName = "Color", required = false, readOnly = false,
-        defaultValue = "#0066CC",
-        description = "Hex color code for priority visualization (e.g., #FF0000 for urgent)",
-        hidden = false, order = 4, maxLength = 7
-    )
-    private String color = "#0066CC";
+	/**
+	 * Indicates if this is the default priority.
+	 */
+	@Column (name = "is_default", nullable = false)
+	@MetaData (
+		displayName = "Is Default", required = false, readOnly = false,
+		defaultValue = "false", description = "Indicates if this is the default priority", 
+		hidden = false, order = 7
+	)
+	private Boolean isDefault = false;
 
-    @Column(name = "is_default", nullable = false)
-    @MetaData(
-        displayName = "Is Default Priority", required = true, readOnly = false,
-        defaultValue = "false",
-        description = "Indicates if this is the default priority for new comments",
-        hidden = false, order = 5
-    )
-    private boolean isDefault = false;
+	/**
+	 * Default constructor for JPA.
+	 */
+	public CCommentPriority() {
+		super();
+		LOGGER.debug("CCommentPriority default constructor called");
+	}
 
-    /**
-     * Default constructor for JPA.
-     */
-    public CCommentPriority() {
-        super();
-    }
+	/**
+	 * Constructor with required fields only.
+	 * @param name    the name of the comment priority (e.g., "URGENT", "NORMAL")
+	 * @param project the project this priority belongs to
+	 */
+	public CCommentPriority(final String name, final CProject project) {
+		super(name, project);
+		LOGGER.debug("CCommentPriority constructor called with name: {} and project: {}",
+			name, project);
+	}
 
-    /**
-     * Constructor with name and priority level.
-     * @param name the name of the comment priority - must not be null
-     * @param priorityLevel the numeric priority level (1=Highest, 4=Lowest) - must not be null
-     */
-    public CCommentPriority(final String name, final Integer priorityLevel) {
-        super(name);
-        
-        if (priorityLevel == null) {
-            LOGGER.warn("CCommentPriority constructor - Priority level parameter is null");
-        }
-        this.priorityLevel = priorityLevel != null ? priorityLevel : 2;
-    }
+	/**
+	 * Constructor with name and priority level.
+	 * @param name the name of the comment priority
+	 * @param priorityLevel the priority level (1=Highest, 4=Lowest)
+	 */
+	public CCommentPriority(final String name, final Integer priorityLevel) {
+		super();
+		setName(name);
+		this.PriorityLevel = priorityLevel;
+		LOGGER.debug("CCommentPriority constructor called with name: {} and priorityLevel: {}",
+			name, priorityLevel);
+	}
 
-    /**
-     * Constructor with name, priority level, and description.
-     * @param name the name of the comment priority - must not be null
-     * @param priorityLevel the numeric priority level (1=Highest, 4=Lowest) - must not be null
-     * @param description description of the comment priority
-     */
-    public CCommentPriority(final String name, final Integer priorityLevel, final String description) {
-        super(name, description);
-        
-        if (priorityLevel == null) {
-            LOGGER.warn("CCommentPriority constructor - Priority level parameter is null");
-        }
-        this.priorityLevel = priorityLevel != null ? priorityLevel : 2;
-    }
+	/**
+	 * Constructor with all common fields.
+	 * @param name      the name of the comment priority
+	 * @param project   the project this priority belongs to
+	 * @param color     the hex color code for UI display
+	 * @param sortOrder the display order
+	 */
+	public CCommentPriority(final String name, final CProject project, final String color,
+		final Integer sortOrder) {
+		super(name, project);
+		setColor(color);
+		setSortOrder(sortOrder);
+		LOGGER.debug(
+			"CCommentPriority constructor called with name: {}, project: {}, color: {}, sortOrder: {}",
+			name, project, color, sortOrder);
+	}
 
-    public String getColor() {
-        return color;
-    }
+	public Integer getPriorityLevel() { return PriorityLevel; }
 
-    public Integer getPriorityLevel() {
-        return priorityLevel;
-    }
+	public void setPriorityLevel(final Integer priorityLevel) {
+		PriorityLevel = priorityLevel;
+	}
 
-    public boolean isDefault() {
-        return isDefault;
-    }
+	/**
+	 * Gets the default status of this priority.
+	 * 
+	 * @return true if this is the default priority, false otherwise
+	 */
+	public Boolean getIsDefault() {
+		return isDefault;
+	}
 
-    public void setColor(final String color) {
-        LOGGER.info("setColor called with color: {}", color);
-        this.color = color;
-        updateLastModified();
-    }
+	/**
+	 * Sets the default status of this priority.
+	 * 
+	 * @param isDefault true to set as default, false otherwise
+	 */
+	public void setDefault(final Boolean isDefault) {
+		this.isDefault = isDefault;
+	}
 
-    public void setDefault(final boolean isDefault) {
-        LOGGER.info("setDefault called with isDefault: {}", isDefault);
-        this.isDefault = isDefault;
-        updateLastModified();
-    }
+	/**
+	 * Convenience method to check if this is the default priority.
+	 * 
+	 * @return true if this is the default priority, false otherwise
+	 */
+	public boolean isDefault() {
+		return Boolean.TRUE.equals(isDefault);
+	}
 
-    public void setPriorityLevel(final Integer priorityLevel) {
-        LOGGER.info("setPriorityLevel called with priorityLevel: {}", priorityLevel);
-        
-        if (priorityLevel == null) {
-            LOGGER.warn("setPriorityLevel called with null priorityLevel");
-            return;
-        }
-        
-        this.priorityLevel = priorityLevel;
-        updateLastModified();
-    }
-
-    @Override
-    protected void initializeDefaults() {
-        super.initializeDefaults();
-        
-        if (this.priorityLevel == null) {
-            this.priorityLevel = 2; // Default to NORMAL priority
-        }
-        
-        if (this.color == null || this.color.trim().isEmpty()) {
-            this.color = "#0066CC"; // Default blue color
-        }
-    }
-
-    @Override
-    public String toString() {
-        return String.format("%s (Level: %d, Color: %s)", getName(), priorityLevel, color);
-    }
+	@Override
+	public String toString() {
+		return String.format(
+			"CCommentPriority{id=%d, name='%s', color='%s', sortOrder=%d, isActive=%s, project=%s, priorityLevel=%d, isDefault=%s}",
+			getId(), getName(), getColor(), getSortOrder(), getIsActive(),
+			getProject() != null ? getProject().getName() : "null", PriorityLevel, isDefault);
+	}
 }
