@@ -7,9 +7,9 @@ import jakarta.persistence.AttributeOverride;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
-import jakarta.validation.constraints.Size;
 import tech.derbent.abstracts.annotations.MetaData;
 import tech.derbent.base.domain.CStatus;
+import tech.derbent.projects.domain.CProject;
 
 /**
  * CDecisionStatus - Domain entity representing decision status types. Layer: Domain (MVC)
@@ -23,16 +23,6 @@ import tech.derbent.base.domain.CStatus;
 public class CDecisionStatus extends CStatus {
 
 	private static final Logger LOGGER = LoggerFactory.getLogger(CDecisionStatus.class);
-
-	@Column (name = "color", nullable = true, length = 7)
-	@Size (max = 7)
-	@MetaData (
-		displayName = "Color", required = false, readOnly = false,
-		defaultValue = "#808080",
-		description = "Hex color code for status visualization (e.g., #FF0000)",
-		hidden = false, order = 3, maxLength = 7
-	)
-	private String color = "#808080";
 
 	@Column (name = "is_final", nullable = false)
 	@MetaData (
@@ -49,16 +39,7 @@ public class CDecisionStatus extends CStatus {
 		defaultValue = "100", description = "Display order for status sorting",
 		hidden = false, order = 5
 	)
-	private Integer sortOrder = 100;
-
-	@Column (name = "allows_editing", nullable = false)
-	@MetaData (
-		displayName = "Allows Editing", required = true, readOnly = false,
-		defaultValue = "true",
-		description = "Whether decisions with this status can be edited", hidden = false,
-		order = 6
-	)
-	private boolean allowsEditing = true;
+	private final Integer sortOrder = 100;
 
 	@Column (name = "requires_approval", nullable = false)
 	@MetaData (
@@ -77,24 +58,12 @@ public class CDecisionStatus extends CStatus {
 	}
 
 	/**
-	 * Constructor with name.
-	 * @param name the name of the decision status - must not be null or empty
-	 */
-	public CDecisionStatus(final String name) {
-		super(name);
-
-		if ((name == null) || name.trim().isEmpty()) {
-			LOGGER.warn("CDecisionStatus constructor - Name parameter is null or empty");
-		}
-	}
-
-	/**
 	 * Constructor with name and description.
 	 * @param name        the name of the decision status - must not be null or empty
 	 * @param description detailed description of the decision status - can be null
 	 */
-	public CDecisionStatus(final String name, final String description) {
-		super(name, description);
+	public CDecisionStatus(final String name, final CProject project) {
+		super(name, project);
 	}
 
 	/**
@@ -106,26 +75,17 @@ public class CDecisionStatus extends CStatus {
 	 * @param allowsEditing    whether decisions with this status can be edited
 	 * @param requiresApproval whether decisions with this status require approval
 	 */
-	public CDecisionStatus(final String name, final String description,
-		final String color, final boolean isFinal, final boolean allowsEditing,
+	public CDecisionStatus(final String name, final CProject project,
+		final String description, final String color, final boolean isFinal,
 		final boolean requiresApproval) {
-		super(name, description);
+		super(name, project);
 		LOGGER.debug(
-			"CDecisionStatus constructor called with name: {}, description: {}, color: {}, isFinal: {}, allowsEditing: {}, requiresApproval: {}",
-			name, description, color, isFinal, allowsEditing, requiresApproval);
-		this.color = color != null ? color : "#808080";
+			"CDecisionStatus constructor called with name: {}, description: {}, color: {}, isFinal: {}, requiresApproval: {}",
+			name, description, color, isFinal, requiresApproval);
+		setColor(color);
+		setDescription(description);
 		this.isFinal = isFinal;
-		this.allowsEditing = allowsEditing;
 		this.requiresApproval = requiresApproval;
-	}
-	// Getters and Setters
-
-	/**
-	 * Checks if decisions with this status can be modified.
-	 * @return true if editing is allowed and status is not final
-	 */
-	public boolean canBeModified() {
-		return allowsEditing && !isFinal;
 	}
 
 	@Override
@@ -140,19 +100,6 @@ public class CDecisionStatus extends CStatus {
 		}
 		return super.equals(o);
 	}
-
-	public String getColor() {
-		return (color != null) && !color.trim().isEmpty() ? color : "#808080";
-	}
-
-	public Integer getSortOrder() { return sortOrder != null ? sortOrder : 100; }
-
-	@Override
-	public int hashCode() {
-		return super.hashCode();
-	}
-
-	public boolean isAllowsEditing() { return allowsEditing; }
 
 	/**
 	 * Checks if this status indicates completion of the decision process.
@@ -169,17 +116,6 @@ public class CDecisionStatus extends CStatus {
 	public boolean isPendingApproval() { return requiresApproval && !isFinal; }
 
 	public boolean isRequiresApproval() { return requiresApproval; }
-	// Business Logic Methods
-
-	public void setAllowsEditing(final boolean allowsEditing) {
-		this.allowsEditing = allowsEditing;
-		updateLastModified();
-	}
-
-	public void setColor(final String color) {
-		this.color = (color != null) && !color.trim().isEmpty() ? color : "#808080";
-		updateLastModified();
-	}
 
 	public void setFinal(final boolean isFinal) {
 		this.isFinal = isFinal;
@@ -188,11 +124,6 @@ public class CDecisionStatus extends CStatus {
 
 	public void setRequiresApproval(final boolean requiresApproval) {
 		this.requiresApproval = requiresApproval;
-		updateLastModified();
-	}
-
-	public void setSortOrder(final Integer sortOrder) {
-		this.sortOrder = sortOrder != null ? sortOrder : 100;
 		updateLastModified();
 	}
 
