@@ -9,226 +9,204 @@ import org.springframework.transaction.annotation.Transactional;
 
 import tech.derbent.abstracts.services.CAbstractNamedEntityService;
 import tech.derbent.decisions.domain.CDecisionStatus;
+import tech.derbent.projects.domain.CProject;
 
 /**
- * CDecisionStatusService - Service class for CDecisionStatus entities.
- * Layer: Service (MVC)
- * 
- * Provides business logic operations for decision status management including
+ * CDecisionStatusService - Service class for CDecisionStatus entities. Layer: Service
+ * (MVC) Provides business logic operations for decision status management including
  * validation, creation, and workflow management.
  */
 @Service
-@PreAuthorize("isAuthenticated()")
+@PreAuthorize ("isAuthenticated()")
 public class CDecisionStatusService extends CAbstractNamedEntityService<CDecisionStatus> {
 
-    public CDecisionStatusService(final CDecisionStatusRepository repository, final Clock clock) {
-        super(repository, clock);
-    }
+	public CDecisionStatusService(final CDecisionStatusRepository repository,
+		final Clock clock) {
+		super(repository, clock);
+	}
 
-    @Override
-    protected CDecisionStatus createNewEntityInstance() {
-        return new CDecisionStatus();
-    }
+	/**
+	 * Creates a new decision status with basic properties.
+	 * @param name        the decision status name - must not be null or empty
+	 * @param description the description - can be null
+	 * @param isFinal     whether this is a final status
+	 * @return the created decision status
+	 */
+	@Transactional
+	public CDecisionStatus createDecisionStatus(final String name, final CProject project,
+		final String description, final boolean isFinal) {
+		LOGGER.info(
+			"createDecisionStatus called with name: {}, description: {}, isFinal: {}",
+			name, description, isFinal);
 
-    /**
-     * Finds all non-final decision statuses ordered by sort order.
-     * @return list of non-final decision statuses sorted by sort order
-     */
-    @Transactional(readOnly = true)
-    public List<CDecisionStatus> findAllActiveOrdered() {
-        LOGGER.info("findAllActiveOrdered called for decision statuses");
-        return ((CDecisionStatusRepository) repository).findByIsFinalFalseOrderBySortOrderAsc();
-    }
+		if ((name == null) || name.trim().isEmpty()) {
+			LOGGER.error("createDecisionStatus called with null or empty name");
+			throw new IllegalArgumentException(
+				"Decision status name cannot be null or empty");
+		}
+		final CDecisionStatus decisionStatus = new CDecisionStatus(name.trim(), project);
+		decisionStatus.setFinal(isFinal);
+		decisionStatus.setDescription(description);
+		return repository.saveAndFlush(decisionStatus);
+	}
 
-    /**
-     * Finds all decision statuses ordered by sort order.
-     * @return list of decision statuses sorted by sort order
-     */
-    @Transactional(readOnly = true)
-    public List<CDecisionStatus> findAllOrdered() {
-        LOGGER.info("findAllOrdered called for decision statuses");
-        return ((CDecisionStatusRepository) repository).findAllByOrderBySortOrderAsc();
-    }
+	/**
+	 * Creates a new decision status with all properties.
+	 * @param name             the decision status name - must not be null or empty
+	 * @param description      the description - can be null
+	 * @param color            the hex color code - can be null
+	 * @param isFinal          whether this is a final status
+	 * @param allowsEditing    whether decisions with this status can be edited
+	 * @param requiresApproval whether decisions with this status require approval
+	 * @return the created decision status
+	 */
+	@Transactional
+	public CDecisionStatus createDecisionStatus(final String name, final CProject project,
+		final String description, final String color, final boolean isFinal,
+		final boolean requiresApproval) {
+		LOGGER.info(
+			"createDecisionStatus called with name: {}, description: {}, color: {}, isFinal: {}, requiresApproval: {}",
+			name, description, color, isFinal, requiresApproval);
 
-    /**
-     * Finds all final decision statuses.
-     * @return list of final decision statuses
-     */
-    @Transactional(readOnly = true)
-    public List<CDecisionStatus> findAllFinal() {
-        LOGGER.info("findAllFinal called for decision statuses");
-        return ((CDecisionStatusRepository) repository).findByIsFinalTrue();
-    }
+		if ((name == null) || name.trim().isEmpty()) {
+			LOGGER.error("createDecisionStatus called with null or empty name");
+			throw new IllegalArgumentException(
+				"Decision status name cannot be null or empty");
+		}
+		final CDecisionStatus decisionStatus = new CDecisionStatus(name.trim(), project,
+			description, color, isFinal, requiresApproval);
+		return repository.saveAndFlush(decisionStatus);
+	}
 
-    /**
-     * Finds decision statuses that allow editing.
-     * @return list of decision statuses that allow editing
-     */
-    @Transactional(readOnly = true)
-    public List<CDecisionStatus> findAllowingEditing() {
-        LOGGER.info("findAllowingEditing called for decision statuses");
-        return ((CDecisionStatusRepository) repository).findByAllowsEditingTrue();
-    }
+	@Override
+	protected CDecisionStatus createNewEntityInstance() {
+		return new CDecisionStatus();
+	}
 
-    /**
-     * Finds decision statuses that require approval.
-     * @return list of decision statuses that require approval
-     */
-    @Transactional(readOnly = true)
-    public List<CDecisionStatus> findRequiringApproval() {
-        LOGGER.info("findRequiringApproval called for decision statuses");
-        return ((CDecisionStatusRepository) repository).findByRequiresApprovalTrue();
-    }
+	/**
+	 * Finds all non-final decision statuses ordered by sort order.
+	 * @return list of non-final decision statuses sorted by sort order
+	 */
+	@Transactional (readOnly = true)
+	public List<CDecisionStatus> findAllActiveOrdered() {
+		LOGGER.info("findAllActiveOrdered called for decision statuses");
+		return ((CDecisionStatusRepository) repository)
+			.findByIsFinalFalseOrderBySortOrderAsc();
+	}
 
-    /**
-     * Creates a new decision status with basic properties.
-     * @param name the decision status name - must not be null or empty
-     * @param description the description - can be null
-     * @param isFinal whether this is a final status
-     * @return the created decision status
-     */
-    @Transactional
-    public CDecisionStatus createDecisionStatus(final String name, final String description, 
-                                              final boolean isFinal) {
-        LOGGER.info("createDecisionStatus called with name: {}, description: {}, isFinal: {}", 
-                   name, description, isFinal);
-        
-        if (name == null || name.trim().isEmpty()) {
-            LOGGER.error("createDecisionStatus called with null or empty name");
-            throw new IllegalArgumentException("Decision status name cannot be null or empty");
-        }
-        
-        final CDecisionStatus decisionStatus = new CDecisionStatus(name.trim(), description);
-        decisionStatus.setFinal(isFinal);
-        
-        return repository.saveAndFlush(decisionStatus);
-    }
+	/**
+	 * Finds all final decision statuses.
+	 * @return list of final decision statuses
+	 */
+	@Transactional (readOnly = true)
+	public List<CDecisionStatus> findAllFinal() {
+		LOGGER.info("findAllFinal called for decision statuses");
+		return ((CDecisionStatusRepository) repository).findByIsFinalTrue();
+	}
 
-    /**
-     * Creates a new decision status with all properties.
-     * @param name the decision status name - must not be null or empty
-     * @param description the description - can be null
-     * @param color the hex color code - can be null
-     * @param isFinal whether this is a final status
-     * @param allowsEditing whether decisions with this status can be edited
-     * @param requiresApproval whether decisions with this status require approval
-     * @return the created decision status
-     */
-    @Transactional
-    public CDecisionStatus createDecisionStatus(final String name, final String description, 
-                                              final String color, final boolean isFinal,
-                                              final boolean allowsEditing, 
-                                              final boolean requiresApproval) {
-        LOGGER.info("createDecisionStatus called with name: {}, description: {}, color: {}, isFinal: {}, allowsEditing: {}, requiresApproval: {}", 
-                   name, description, color, isFinal, allowsEditing, requiresApproval);
-        
-        if (name == null || name.trim().isEmpty()) {
-            LOGGER.error("createDecisionStatus called with null or empty name");
-            throw new IllegalArgumentException("Decision status name cannot be null or empty");
-        }
-        
-        final CDecisionStatus decisionStatus = new CDecisionStatus(name.trim(), description, color, 
-                                                                 isFinal, allowsEditing, requiresApproval);
-        
-        return repository.saveAndFlush(decisionStatus);
-    }
+	/**
+	 * Finds all decision statuses ordered by sort order.
+	 * @return list of decision statuses sorted by sort order
+	 */
+	@Transactional (readOnly = true)
+	public List<CDecisionStatus> findAllOrdered() {
+		LOGGER.info("findAllOrdered called for decision statuses");
+		return ((CDecisionStatusRepository) repository).findAllByOrderBySortOrderAsc();
+	}
 
-    /**
-     * Updates the sort order of a decision status.
-     * @param decisionStatus the decision status to update - must not be null
-     * @param sortOrder the new sort order
-     * @return the updated decision status
-     */
-    @Transactional
-    public CDecisionStatus updateSortOrder(final CDecisionStatus decisionStatus, final Integer sortOrder) {
-        LOGGER.info("updateSortOrder called with decisionStatus: {}, sortOrder: {}", 
-                   decisionStatus != null ? decisionStatus.getName() : "null", sortOrder);
-        
-        if (decisionStatus == null) {
-            LOGGER.error("updateSortOrder called with null decisionStatus");
-            throw new IllegalArgumentException("Decision status cannot be null");
-        }
-        
-        decisionStatus.setSortOrder(sortOrder);
-        return repository.saveAndFlush(decisionStatus);
-    }
+	/**
+	 * Finds decision statuses that allow editing.
+	 * @return list of decision statuses that allow editing
+	 */
+	@Transactional (readOnly = true)
+	public List<CDecisionStatus> findAllowingEditing() {
+		LOGGER.info("findAllowingEditing called for decision statuses");
+		return ((CDecisionStatusRepository) repository).findByAllowsEditingTrue();
+	}
 
-    /**
-     * Updates the properties of a decision status.
-     * @param decisionStatus the decision status to update - must not be null
-     * @param allowsEditing whether decisions with this status can be edited
-     * @param requiresApproval whether decisions with this status require approval
-     * @return the updated decision status
-     */
-    @Transactional
-    public CDecisionStatus updateStatusProperties(final CDecisionStatus decisionStatus, 
-                                                final boolean allowsEditing, 
-                                                final boolean requiresApproval) {
-        LOGGER.info("updateStatusProperties called with decisionStatus: {}, allowsEditing: {}, requiresApproval: {}", 
-                   decisionStatus != null ? decisionStatus.getName() : "null", allowsEditing, requiresApproval);
-        
-        if (decisionStatus == null) {
-            LOGGER.error("updateStatusProperties called with null decisionStatus");
-            throw new IllegalArgumentException("Decision status cannot be null");
-        }
-        
-        decisionStatus.setAllowsEditing(allowsEditing);
-        decisionStatus.setRequiresApproval(requiresApproval);
-        
-        return repository.saveAndFlush(decisionStatus);
-    }
+	/**
+	 * Finds decision statuses that require approval.
+	 * @return list of decision statuses that require approval
+	 */
+	@Transactional (readOnly = true)
+	public List<CDecisionStatus> findRequiringApproval() {
+		LOGGER.info("findRequiringApproval called for decision statuses");
+		return ((CDecisionStatusRepository) repository).findByRequiresApprovalTrue();
+	}
 
-    /**
-     * Checks if decisions with this status can be modified.
-     * @param decisionStatus the decision status to check
-     * @return true if editing is allowed and status is not final
-     */
-    @Transactional(readOnly = true)
-    public boolean canDecisionsBeModified(final CDecisionStatus decisionStatus) {
-        LOGGER.info("canDecisionsBeModified called with decisionStatus: {}", 
-                   decisionStatus != null ? decisionStatus.getName() : "null");
-        
-        if (decisionStatus == null) {
-            LOGGER.warn("canDecisionsBeModified called with null decisionStatus");
-            return false;
-        }
-        
-        return decisionStatus.canBeModified();
-    }
+	/**
+	 * Checks if this status indicates completion of the decision process.
+	 * @param decisionStatus the decision status to check
+	 * @return true if this is a final status
+	 */
+	@Transactional (readOnly = true)
+	public boolean isStatusCompleted(final CDecisionStatus decisionStatus) {
+		LOGGER.info("isStatusCompleted called with decisionStatus: {}",
+			decisionStatus != null ? decisionStatus.getName() : "null");
 
-    /**
-     * Checks if this status indicates completion of the decision process.
-     * @param decisionStatus the decision status to check
-     * @return true if this is a final status
-     */
-    @Transactional(readOnly = true)
-    public boolean isStatusCompleted(final CDecisionStatus decisionStatus) {
-        LOGGER.info("isStatusCompleted called with decisionStatus: {}", 
-                   decisionStatus != null ? decisionStatus.getName() : "null");
-        
-        if (decisionStatus == null) {
-            LOGGER.warn("isStatusCompleted called with null decisionStatus");
-            return false;
-        }
-        
-        return decisionStatus.isCompleted();
-    }
+		if (decisionStatus == null) {
+			LOGGER.warn("isStatusCompleted called with null decisionStatus");
+			return false;
+		}
+		return decisionStatus.isCompleted();
+	}
 
-    /**
-     * Checks if decisions with this status are pending approval.
-     * @param decisionStatus the decision status to check
-     * @return true if approval is required and status is not final
-     */
-    @Transactional(readOnly = true)
-    public boolean isStatusPendingApproval(final CDecisionStatus decisionStatus) {
-        LOGGER.info("isStatusPendingApproval called with decisionStatus: {}", 
-                   decisionStatus != null ? decisionStatus.getName() : "null");
-        
-        if (decisionStatus == null) {
-            LOGGER.warn("isStatusPendingApproval called with null decisionStatus");
-            return false;
-        }
-        
-        return decisionStatus.isPendingApproval();
-    }
+	/**
+	 * Checks if decisions with this status are pending approval.
+	 * @param decisionStatus the decision status to check
+	 * @return true if approval is required and status is not final
+	 */
+	@Transactional (readOnly = true)
+	public boolean isStatusPendingApproval(final CDecisionStatus decisionStatus) {
+		LOGGER.info("isStatusPendingApproval called with decisionStatus: {}",
+			decisionStatus != null ? decisionStatus.getName() : "null");
+
+		if (decisionStatus == null) {
+			LOGGER.warn("isStatusPendingApproval called with null decisionStatus");
+			return false;
+		}
+		return decisionStatus.isPendingApproval();
+	}
+
+	/**
+	 * Updates the sort order of a decision status.
+	 * @param decisionStatus the decision status to update - must not be null
+	 * @param sortOrder      the new sort order
+	 * @return the updated decision status
+	 */
+	@Transactional
+	public CDecisionStatus updateSortOrder(final CDecisionStatus decisionStatus,
+		final Integer sortOrder) {
+		LOGGER.info("updateSortOrder called with decisionStatus: {}, sortOrder: {}",
+			decisionStatus != null ? decisionStatus.getName() : "null", sortOrder);
+
+		if (decisionStatus == null) {
+			LOGGER.error("updateSortOrder called with null decisionStatus");
+			throw new IllegalArgumentException("Decision status cannot be null");
+		}
+		decisionStatus.setSortOrder(sortOrder);
+		return repository.saveAndFlush(decisionStatus);
+	}
+
+	/**
+	 * Updates the properties of a decision status.
+	 * @param decisionStatus   the decision status to update - must not be null
+	 * @param allowsEditing    whether decisions with this status can be edited
+	 * @param requiresApproval whether decisions with this status require approval
+	 * @return the updated decision status
+	 */
+	@Transactional
+	public CDecisionStatus updateStatusProperties(final CDecisionStatus decisionStatus,
+		final boolean requiresApproval) {
+		LOGGER.info(
+			"updateStatusProperties called with decisionStatus: {}, requiresApproval: {}",
+			decisionStatus != null ? decisionStatus.getName() : "null", requiresApproval);
+
+		if (decisionStatus == null) {
+			LOGGER.error("updateStatusProperties called with null decisionStatus");
+			throw new IllegalArgumentException("Decision status cannot be null");
+		}
+		decisionStatus.setRequiresApproval(requiresApproval);
+		return repository.saveAndFlush(decisionStatus);
+	}
 }
