@@ -7,7 +7,6 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import org.junit.jupiter.api.Test;
 
 import tech.derbent.activities.domain.CActivity;
-import tech.derbent.activities.domain.CActivityStatus;
 import tech.derbent.projects.domain.CProject;
 
 /**
@@ -43,25 +42,9 @@ class CActivityCardTest {
 
 	@Test
 	void testActivityCardWithNullThrowsException() {
-		// Given - null activity
-		// When & Then
+		// Given - null activity When & Then
 		assertThrows(IllegalArgumentException.class, () -> {
 			new CActivityCard(null);
 		});
-	}
-
-	@Test
-	void testActivityCardWithStatus() {
-		// Given
-		final CProject project = new CProject();
-		project.setName("Test Project");
-		final CActivity activity = new CActivity("Test Activity", project);
-		final CActivityStatus status = new CActivityStatus("In Progress");
-		activity.setStatus(status);
-		// When
-		final CActivityCard card = new CActivityCard(activity);
-		// Then
-		assertNotNull(card);
-		assertEquals(activity, card.getActivity());
 	}
 }
